requirements-airflow.txt
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-redis==3.2.0
apache-airflow-providers-celery==3.5.0
celery>=5.3.0
redis>=4.5.0
python-dotenv>=1.0.1

# Pipeline data
pandas==2.2.2
geopandas==0.14.4
pyarrow==16.1.0
py7zr==0.21.0
shapely==2.0.4
numpy==1.26.4
requests==2.32.3
openpyxl==3.1.2
pyogrio>=0.7.0
sqlalchemy>=1.4.28,<2.0
psycopg2-binary>=2.9.0
pymongo>=4.8.0